Refactor relay routes to relay-route module
