+++ app/component/map/tile-layer/TileLayerContainer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Relay from 'react-relay';
-import StopRoute from './StopRoute';
-import CityBikeRoute from './CityBikeRoute';
+import StopRoute from '../../../relay-route/StopRoute';
+import CityBikeRoute from '../../../relay-route/CityBikeRoute';
 import Popup from '../dynamic-popup';
 import intl from 'react-intl';
 import BaseTileLayer from 'react-leaflet/lib/BaseTileLayer';
